refactor: use vanilla-extract in tooltip component
